interaction_service/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
neo4j>=5.20,<6.0
